angkutan_tbs_template/requirements.txt

streamlit
pandas
streamlit-authenticator
pyyaml
